requirements.txt
asgiref
certifi
cloudinary
dj-database-url
Django
django-cors-headers
django-filter
djangorestframework
gunicorn
Pillow
psycopg2
psycopg2-binary
pytz
six
sqlparse
urllib3
whitenoise